Add files via upload
(Using R software) Task 
The data can be retrieved from the following link.
https://public.tableau.com/app/sample-data/HollywoodsMostProfitableStories.csv
Steps involved (Storytelling):
The raw data was downloaded and ready to go through for the process. R software is useful for data analysing and it’s precise tuning for better presentation. User can see it in better format with lowest chance of data being wrongly analysed.